wnba-pulse-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic==2.9.2
APScheduler==3.10.4
redis==5.0.8
requests==2.32.3
beautifulsoup4==4.12.3
praw==7.7.1
vaderSentiment==3.3.2
python-dotenv==1.0.1
structlog==24.1.0
